% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplr.R
\name{nplr}
\alias{nplr}
\title{Function to Fit n-Parameter Logistic Regressions.}
\usage{
nplr(
  x,
  y,
  useLog = TRUE,
  LPweight = 0.25,
  npars = "all",
  method = c("res", "sdw", "gw"),
  silent = FALSE
)
}
\arguments{
\item{x}{a vector of numeric values, e.g. a vector of drug concentrations.}

\item{y}{a vector of numeric values, e.g. a vector of responses, typically provided as proportions of control.}

\item{useLog}{Logical. Should x-values be Log10-transformed. Default to \code{TRUE}, set to \code{FALSE} if x is already in Log10.}

\item{LPweight}{a coefficient to adjust the weights. \eqn{LPweight = 0} will compute a non-weighted np-logistic regression.}

\item{npars}{a numeric value (or \code{"all"}) to specify the number of parameters to use in the model. If \code{"all"} the logistic model will be tested with 2 to 5 parameters, and the best option will be returned. See Details.}

\item{method}{a character string to specify what weight method to use. Options are \code{"res" (Default), "sdw", "gw"}. See Details.}

\item{silent}{Logical. Specify whether \code{warnings} and/or \code{messages} have to be silenced. Default to \code{FALSE}.}
}
\value{
An object of class \code{nplr}.
}
\description{
This function computes a weighted n-parameters logistic regression, given x (typically compound concentrations) and y values (responses: optic densities, fluorescence, cell counts,...). See Details.
}
\details{
The 5-parameter logistic regression is of the form:

\deqn{ y = B + (T - B)/[1 + 10^(b*(xmid - x))]^s }

where \code{B} and \code{T} are the bottom and top asymptotes, respectively, \code{b} and \code{xmid} are the Hill slope and the x-coordinate at the inflexion point, respectively, and \code{s} is an asymmetric coefficient. This equation is sometimes referred to as the Richards' equation.

When specifying \code{npars = 4}, the \code{s} parameter is forced to be \code{1}, and the corresponding model is a 4-parameter logistic regression, symmetrical around its inflexion point. When specifying \code{npars = 3} or \code{npars = 2}, additional constraints force \code{B} and \code{T} to be \code{0} and \code{1}, respectively.

Weight methods:

The model parameters are optimized using \code{\link[stats]{nlm}}, given a sum of squared errors function, \eqn{sse(Y)}, to minimize:

\deqn{sse(Y) = \Sigma [W.(Yobs - Yfit)^2 ]}

where \eqn{Yobs}, \eqn{Yfit}, and \eqn{W} are the vectors of observed values, fitted values, and weights, respectively.

- Residual weights (\code{"res"}): \deqn{W = (1/residuals)^LPweight}, where \code{residuals} are the squared error between observed and fitted values, and \code{LPweight} is a tuning parameter.
- Standard weights (\code{"sdw"}): \deqn{W = 1/Var(Yobs_r)}, where \code{Var(Yobs_r)} is the vector of within-replicates variances.
- General weights (\code{"gw"}): \deqn{W = 1/Yfit^LPweight}, where \code{Yfit} are the fitted values.

The standard and general weight methods are described in the references.
}
\note{
The data used in the examples are samples from the NCI-60 Growth Inhibition Data: \url{https://wiki.nci.nih.gov/display/NCIDTPdata/NCI-60+Growth+Inhibition+Data}, except for multicell.tsv which are simulated data.
}
\section{Slots}{

\itemize{
  \item \code{x}: the x values used in the model, which can be \code{Log10(x)} if \code{useLog = TRUE}.
  \item \code{y}: the y values.
  \item \code{useLog}: logical.
  \item \code{npars}: the best number of parameters if \code{npars = "all"}, or the specified number otherwise.
  \item \code{LPweight}: the weights tuning parameter.
  \item \code{yFit}: the y fitted values.
  \item \code{xCurve}: the x values generated to draw the curve (200 points between \code{min} and \code{max} of x).
  \item \code{yCurve}: the fitted values corresponding to \code{xCurve}.
  \item \code{inflPoint}: the inflexion point x and y coordinates.
  \item \code{goodness}: the goodness-of-fit (correlation between fitted and observed y values).
  \item \code{stdErr}: the mean squared error between fitted and observed y values.
  \item \code{pars}: the model parameters.
  \item \code{AUC}: the area under the curve, estimated using both the trapezoid method and Simpson's rule.
}
}

\examples{
require(nplr)
path <- system.file("extdata", "pc3.txt", package = "nplr")
pc3 <- read.delim(path)
model <- nplr(x = pc3$CONC, y = pc3$GIPROP)
plot(model)
}
\references{
1. Richards, F. J. (1959). A flexible growth function for empirical use. J Exp Bot 10, 290-300.
2. Giraldo J, Vivas NM, Vila E, Badia A. Assessing the (a)symmetry of concentration-effect curves: empirical versus mechanistic models. Pharmacol Ther. 2002 Jul;95(1):21-45.
3. Motulsky HJ, Brown RE. Detecting outliers when fitting data with nonlinear regression - a new method based on robust nonlinear regression and the false discovery rate. BMC Bioinformatics. 2006 Mar 9;7:123.
}
\seealso{
\code{\link{convertToProp}}, \code{\link{getEstimates}}, \code{\link{plot.nplr}}, \code{\link{nplrAccessors}}
}
\author{
Frederic Commo, Brian M. Bot
}
