% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{nplrAccessors}
\alias{nplrAccessors}
\alias{getX}
\alias{getX-methods}
\alias{getX,nplr-method}
\alias{getY}
\alias{getY-methods}
\alias{getY,nplr-method}
\alias{getXcurve}
\alias{getXcurve-methods}
\alias{getXcurve,nplr-method}
\alias{getYcurve}
\alias{getYcurve-methods}
\alias{getYcurve,nplr-method}
\alias{getFitValues}
\alias{getFitValues-methods}
\alias{getFitValues,nplr-method}
\alias{getInflexion}
\alias{getInflexion-methods}
\alias{getInflexion,nplr-method}
\alias{getPar}
\alias{getPar-methods}
\alias{getPar,nplr-method}
\alias{getAUC}
\alias{getAUC-methods}
\alias{getAUC,nplr-method}
\alias{getGoodness}
\alias{getGoodness-methods}
\alias{getGoodness,nplr-method}
\alias{getStdErr}
\alias{getStdErr-methods}
\alias{getStdErr,nplr-method}
\alias{getWeights}
\alias{getWeights-methods}
\alias{getWeights,nplr-method}
\alias{nplrAccessors-methods}
\alias{nplrAccessors,nplr-method}
\title{Accessor Functions for \code{\link{nplr}} Objects}
\usage{
getX(object)
}
\description{
Methods for extracting information from an object of class \code{nplr}. Each of the methods below is a convenience function that extracts the corresponding slot (as suggested by the method name) from the object of class \code{\link{nplr}}.
}
\section{Methods}{

\describe{
  \item{\code{signature(object = "nplr")}}{
  \itemize{
    \item \code{getX(object)}: Extracts the x-values.
    \item \code{getY(object)}: Extracts the y-values.
    \item \code{getXcurve(object)}: Extracts the x-values of the fitted curve.
    \item \code{getYcurve(object)}: Extracts the y-values of the fitted curve.
    \item \code{getFitValues(object)}: Extracts the fitted values.
    \item \code{getInflexion(object)}: Extracts the inflexion point.
    \item \code{getPar(object)}: Extracts the parameters of the logistic model.
    \item \code{getAUC(object)}: Extracts the area under the curve (AUC).
    \item \code{getGoodness(object)}: Extracts the goodness-of-fit.
    \item \code{getStdErr(object)}: Extracts the standard error of the model.
    \item \code{getWeights(object)}: Extracts the weights of the model.
  }
  }
}
}

\seealso{
\code{\link{nplr}}, \code{\link{getEstimates}}
}
