% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nplr.R
\name{summary.nplr}
\alias{summary.nplr}
\title{Summarizing \code{nplr} Objects}
\usage{
\method{summary}{nplr}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{nplr}}.}

\item{...}{Other optional parameters (not used).}
}
\description{
A S3 method to visualize a model summary as a table.
}
\details{
None
}
\examples{
# Using the PC-3 data
require(nplr)
path <- system.file("extdata", "pc3.txt", package = "nplr")
pc3 <- read.delim(path)
model <- nplr(x = pc3$CONC, y = pc3$GIPROP)
summary(model)

}
\seealso{
\code{\link{plot.nplr}}
}
